+++ blog-hub/src/main/java/com/blog_hub/user/service/UserService.java
@@ -1,10 +1,19 @@
 package com.blog_hub.user.service;
 
+import com.blog_hub.post.dto.PostResponse;
+import com.blog_hub.post.entity.Post;
+import com.blog_hub.post.mapper.PostMapper;
+import com.blog_hub.user.dto.CreateUserRequest;
+import com.blog_hub.user.dto.UpdateUserRequest;
+import com.blog_hub.user.dto.UserResponse;
 import com.blog_hub.user.entity.User;
+import com.blog_hub.user.mapper.UserMapper;
 import com.blog_hub.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -12,20 +21,70 @@ public class UserService {
 
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    UserMapper userMapper;
+    @Autowired
+    PostMapper postMapper;
 
-    public List<User> getAllUser(){
-        return userRepository.findAll();
+
+    public List<UserResponse> getAllUser(){
+        List<User> users = userRepository.findAll();
+        List<UserResponse> responses = new ArrayList<>();
+
+        for (User user: users){
+            responses.add(userMapper.toResponse(user));
+        }
+        return responses;
+//        return users.stream()
+//                .map(userMapper::toResponse)
+//                .toList();
     }
 
-    public User getUserById(int id) {
-        return userRepository.getById(id);
+    public UserResponse getUserById(int id) {
+        User user = userRepository.getById(id);
+        return userMapper.toResponse(user);
     }
 
-    public User saveUser(User user) {
-        return userRepository.save(user);
+    public List<PostResponse> getPostsByUser(int id) {
+
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+
+        List<PostResponse> responses = new ArrayList<>();
+        for (Post post : user.getPosts()) {
+            responses.add(postMapper.toResponse(post));
+        }
+
+        return responses;
+//        return user.getPosts()
+//                .stream()
+//                .map(postMapper::toResponse)
+//                .toList();
     }
 
-    public void deleteUser(int id) {
+    public UserResponse saveUser(CreateUserRequest request) {
+        User user = userMapper.toEntity(request);
+        User savedUser = userRepository.save(user);
+
+        return userMapper.toResponse(savedUser);
+    }
+
+    public UserResponse updateUser(int id, UpdateUserRequest request){
+        User user = userRepository.getById(id);
+
+        userMapper.updateUserFromDto(request, user);
+
+        User updatedUser = userRepository.save(user);
+
+        return userMapper.toResponse(updatedUser);
+    }
+
+    public ResponseEntity<?> deleteUser(int id) {
+        User user = userRepository.getById(id);
+        if(user == null){
+            return ResponseEntity.notFound().build();
+        }
         userRepository.deleteById(id);
+        return ResponseEntity.ok(user);
     }
 }
